refactor(doc-history): export makeSemaphore + CPU formula via side-effect-free subpath
Add `./concurrency` subpath to `@takazudo/zudo-doc-history-server` exporting
`makeSemaphore` and `defaultGitConcurrency`. Import from there in
`packages/zudo-doc` pre-build and in cli.ts; delete the verbatim copy from
pre-build.ts. Update semaphore.test.ts import accordingly.

Closes #2028.
