+++ packages/doc-history-server/package.json
@@ -41,6 +41,10 @@
       "types": "./dist/types.d.ts",
       "default": "./dist/types.js"
     },
+    "./concurrency": {
+      "types": "./dist/concurrency.d.ts",
+      "default": "./dist/concurrency.js"
+    },
     "./package.json": "./package.json"
   },
   "files": [
@@ -48,12 +52,12 @@
     "README.md"
   ],
   "scripts": {
-    "build": "tsup src/index.ts src/cli.ts src/git-history.ts src/types.ts --format esm --dts",
+    "build": "tsup src/index.ts src/cli.ts src/git-history.ts src/types.ts src/concurrency.ts --format esm --dts",
     "dev": "tsx src/index.ts",
     "generate": "tsx src/cli.ts",
     "test": "vitest run",
     "typecheck": "tsc --noEmit",
-    "prepare": "tsup src/index.ts src/cli.ts src/git-history.ts src/types.ts --format esm --dts --silent"
+    "prepare": "tsup src/index.ts src/cli.ts src/git-history.ts src/types.ts src/concurrency.ts --format esm --dts --silent"
   },
   "devDependencies": {
     "@types/node": "^22.0.0",
